dist: add dracut module for plymouth-tpm2-totp
Since most distributions using dracut will have plymouth installed as well, we
concentrate on plymouth for now. For user convenience the module is
automatically included in the initramfs if "hostonly" mode is used and a
TOTP secret is stored (which is checked by running "tpm2-totp
calculate"), otherwise the user can still add it manually to
"dracutmodules".

Signed-off-by: Jonas Witschel <diabonas@gmx.de>