tpm2-totp: check if str is NULL in parse_pcrs
When enabling -O2, scan-build on Ubuntu 16.04 generates the warning

/usr/include/x86_64-linux-gnu/bits/string2.h:1154:10: warning: Dereference of null pointer (loaded from variable '__s')
  while (*__s == __sep)
         ^~~~

This is a false positive because the str parameter is never NULL when it
is called in parse_opts, but adding an explicit check does not hurt.

A similar problem has already been addressed in
8d99000daa220b92de11e801d19eee2f3839e462 by initialising saveptr. Since
we now guarantee that str is non-NULL, this initialisiation is not
necessary any more.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>