dist: install soname version of libtss2-tcti-device
Until https://github.com/tpm2-software/tpm2-tss/commit/0638906e2df27431bdfe321aa35ef5d83cceade0,
tpm2-tss was hardcoded to open libtss2-tcti-device.so. For the reasons
explained in the commit message of that commit,
libtss2-tcti-device.so.0 is tried instead now, so we need to make sure
that the soname version of the library is installed to the initramfs.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>