dist: improve plymouth-less mkinitcpio hook
Instead of blocking the initramfs until the user has confirmed the TOTP,
we can run a loop in the background that outputs it every time the TOTP
changes. This way verifying the TOTP and typing in the disk encryption
password can be done in parallel.

Since we do not use any non-POSIX options in the hook any more, the
integration test can be removed.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>