build: install dracut module if plymouth is not available
Since commit 38ca58575fc7b964e60e30457859020df2ba0fe6 the dracut module
works without plymouth as well, so install it whenever dracut is
available.

plymouth-tpm2-totp might not be available even if plymouth is installed
if tpm2-totp was built on a system without plymouth. Fall back to the
version without plymouth in this case, which is guaranteed to exist if
the dracut module is installed. In this case the TOTP is not shown in
the graphical boot screen, but can be seen after pressing the Esc key.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>